move data slice to folder called slices in the feature dir per redux tk best practices
