+++ src/features/slices/dataSlice.ts
@@ -1,0 +1,41 @@
+import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
+import axios from "axios";
+
+interface DataState {
+  data: any;
+  status: "idle" | "loading" | "succeeded" | "failed";
+  error: string | undefined;
+}
+
+const initialState: DataState = {
+  data: null,
+  status: "idle",
+  error: null,
+};
+
+export const fetchData = createAsyncThunk("data/fetchData", async () => {
+  const response = await axios.get("./public/data.json");
+  return response.data;
+});
+
+const dataSlice = createSlice({
+  name: "data",
+  initialState,
+  reducers: {},
+  extraReducers: (builder) => {
+    builder
+      .addCase(fetchData.pending, (state) => {
+        state.status = "loading";
+      })
+      .addCase(fetchData.fulfilled, (state, action) => {
+        state.status = "succeeded";
+        state.data = action.payload;
+      })
+      .addCase(fetchData.rejected, (state, action) => {
+        state.status = "failed";
+        state.error = action.error.message;
+      });
+  },
+});
+
+export default dataSlice.reducer;
